% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMPURE}
\alias{AMPURE}
\title{AMPURE}
\format{
A tibble/data.frame with X rows and Y columns:
\describe{
  \item{Column1}{Description of column1}
  \item{Column2}{Description of column2}
  ...
}
}
\source{
Generated by the Author
}
\usage{
AMPURE
}
\description{
Description: Placeholder for AMPURE dataset.
}
\keyword{datasets}
