% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASV_table}
\alias{ASV_table}
\title{ASV_table}
\format{
A data frame with 9 rows and 3 variables:
\describe{
  \item{sample_name}{Character vector, last digit includes PCR replicate}
  \item{Hash}{The sequence found, hashed using sha1 for consistency}
  \item{nReads}{Number of times that hash was found in that sample}
}
}
\source{
Generated by the Author
}
\usage{
ASV_table
}
\description{
An abundance table in long format.
}
\details{
Contains the number of reads from unique ASVs across different samples.
}
\keyword{datasets}
