% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PCR.R
\name{.tidy_PCR_blocks}
\alias{.tidy_PCR_blocks}
\title{Internal: Tidy PCR sample blocks}
\usage{
.tidy_PCR_blocks(
  df,
  keep = c(Well = 1, Sample = 2, Success = 5, Notes = 6),
  block_size = 6
)
}
\arguments{
\item{df}{A tibble as read from a Google Sheet.}

\item{keep}{Which columns to keep/rename. Default is for normal PCR sheets.
Should be a named character vector like 
`c("Well"=1, "Sample"=2, "Success"=5, "Notes"=6)`.}

\item{block_size}{Number of columns per block (default = 6).}
}
\value{
A tidy tibble.
}
\description{
Helper that takes a messy dataframe with repeated blocks of 6 columns
and converts them into a single tidy tibble with expected columns.
}
\keyword{internal}
