% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Index_PCR}
\alias{Index_PCR}
\title{Index_PCR}
\format{
A character vector
}
\source{
Generated by the Author, stored in googledrive
}
\usage{
Index_PCR
}
\description{
Placeholder description for IndexPCR template web address.
}
\keyword{datasets}
