% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{metadata}
\format{
A data frame with X rows and Y columns:
\describe{
  \item{sample_name}{Name of sample}
  \item{Treatment}{either control or treatment}
   \item{Day}{either 1 or 2}
}
}
\source{
Generated by the Author
}
\usage{
metadata
}
\description{
Placeholder description for metadata dataset.
}
\keyword{datasets}
