% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molarity.data}
\alias{molarity.data}
\title{molarity.data}
\format{
A tibble/data.frame with X rows and Y columns
\describe{
  \item{Sample}{Name of sample}
  \item{Molarity}{in nM}
   \item{Amp_len}{length of amplicon in bp}
   \item{mass}{in ng/\eqn{\mu}l }
}
}
\source{
Generated by the Author
}
\usage{
molarity.data
}
\description{
Placeholder description for molarity.data dataset.
}
\keyword{datasets}
