% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molarityconversions.R
\name{ng2nM}
\alias{ng2nM}
\alias{nM2ng}
\title{functions to translate mass into molarity and vice versa, given we are talking about double stranded DNA
it requires two inputs, the mass (or molarity) and the length of the DNA fragment
It works with the two most common concentrations used in Molecular Ecology labs
 ng/ul for mass
 nM for molarity}
\usage{
ng2nM(ng, length_amplicon)

nM2ng(nM, length_amplicon)
}
\arguments{
\item{ng}{Numeric. the concentration in ng per μL}

\item{length_amplicon}{Integer. The length of the DNA fragment in base pairs.}

\item{nM}{Numeric. The concentration in nmoles per litre}
}
\value{
Numeric. The equivalent concentration in nmoles per litre

Numeric. The equivalent concentration in ng per μL
}
\description{
functions to translate mass into molarity and vice versa, given we are talking about double stranded DNA
it requires two inputs, the mass (or molarity) and the length of the DNA fragment
It works with the two most common concentrations used in Molecular Ecology labs
 ng/ul for mass
 nM for molarity
}
\examples{
\dontrun{
data("molarity.data")
molarity.data |> mutate(moles = ng2nM(ng=mass, length_amplicon = Amp_len))
}
\dontrun{
data("molarity.data")
molarity.data |> mutate(mass = nM2ng(nM=Molarity, length_amplicon = Amp_len))
}

}
