% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OTU_taxonomy}
\alias{OTU_taxonomy}
\title{OTU_taxonomy}
\format{
A dataframe of DNA sequences names and their taxonomy, with 3 rows and 7 columns
 \describe{
  \item{Hash}{Sequence identifier}
  \item{Kingdom}{taxon info at that rank}
   \item{Phylum}{taxon info at that rank}
   \item{Class}{taxon info at that rank}
   \item{Order}{taxon info at that rank}
   \item{Family}{taxon info at that rank}
   \item{Genus}{taxon info at that rank}
   }
}
\source{
Generated by the Author
}
\usage{
OTU_taxonomy
}
\description{
Placeholder description for OTU_taxonomy dataset.
}
\keyword{datasets}
