% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ps}
\alias{ps}
\title{ps}
\format{
Formal class 'phyloseq' with five slots
}
\source{
Generated by the Author
}
\usage{
ps
}
\description{
Placeholder description for phyloseq object.
}
\keyword{datasets}
