% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_PCR}
\alias{template_PCR}
\title{template_PCR}
\format{
A character vector
}
\source{
Generated by the Author, stored in googledrive
}
\usage{
template_PCR
}
\description{
Placeholder description for Normal PCR template web address.
}
\keyword{datasets}
