% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_seqs}
\alias{test_seqs}
\title{test_seqs}
\format{
A dataframe of DNA sequences names and their taxonomy
 \describe{
  \item{Hash}{Sequence identifier}
  \item{Kingdom}{taxon info at that rank}
   \item{Phylum}{taxon info at that rank}
   \item{Class}{taxon info at that rank}
   \item{Order}{taxon info at that rank}
   \item{Family}{taxon info at that rank}
   \item{Genus}{taxon info at that rank}
   
}
}
\source{
Generated by the Author
}
\usage{
test_seqs
}
\description{
Placeholder description for test_seqs dataset.
}
\keyword{datasets}
