% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training.metadata}
\alias{training.metadata}
\title{training.metadata}
\format{
A data frame with 95 rows and 8 columns
\describe{
  \item{Sample_name}{Character vector, last digit includes PCR replicate}
  \item{eDNA.sample}{Biological sample}
  \item{rep}{PCR replicate}
  \item{Transect}{Transect of origin}
   \item{position}{Stop in Transect}
   \item{depth}{in the water column}
   \item{lat}{latitude in decimal degrees}
   \item{lon}{longitude in decimal degrees}
}
}
\source{
Generated by the Author
}
\usage{
training.metadata
}
\description{
Placeholder description for training.metadata dataset.
}
\keyword{datasets}
