% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{tree}
\format{
An object of class phylo or similar
}
\source{
Generated by the Author
}
\usage{
tree
}
\description{
Placeholder description for tree dataset.
}
\keyword{datasets}
