% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UDI_Indices}
\alias{UDI_Indices}
\title{UDI_Indices}
\format{
A data frame with 384 rows and 6 columns
\describe{
  \item{Index}{Character vector}
  \item{Bases_in_Adapter_i7}{Seqs in i7}
  \item{Bases_for_Sample_Sheet_i7}{Seqs to include}
  \item{Bases_for_Sample_Sheet_i5}{Seqs to include in some illumina platoforms}
   \item{Bases_for_Sample_Sheet_i5_B}{Seqs to include in other illumina platoforms}
   \item{Set}{A, B, C, or D}
}
}
\source{
Generated by the Author, data from Illumina
}
\usage{
UDI_Indices
}
\description{
Placeholder description for UDI_Indices dataset.
}
\keyword{datasets}
