% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molarityconversions.R
\name{ng2nM}
\alias{ng2nM}
\title{functions to translate mass into molarity and viceversa, given we are talking about double stranded DNA
it requires two inputs, the mass (or molarity) and the length of the amplicon
It works with the two most common concentrations used in Molecular Ecology labs
 ng/ul for mass
 nM for molarity}
\usage{
ng2nM(ng, length_amplicon)
}
\arguments{
\item{ng}{Numeric. the concentration in ng per uL}

\item{length_amplicon}{Integer. The length of the amplicon in bp.}
}
\value{
Numeric. The equivalent concentration in nmoles per Litre
}
\description{
functions to translate mass into molarity and viceversa, given we are talking about double stranded DNA
it requires two inputs, the mass (or molarity) and the length of the amplicon
It works with the two most common concentrations used in Molecular Ecology labs
 ng/ul for mass
 nM for molarity
}
\examples{
data("molarity.data")
molarity.data \%>\% mutate(moles = ng2nM(ng=mass, length_amplicon = Amp_len))
}
