% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OTU_taxonomy}
\alias{OTU_taxonomy}
\title{OTU_taxonomy}
\format{
A data frame or matrix with X rows and Y columns
}
\source{
Generated by the Author
}
\usage{
OTU_taxonomy
}
\description{
Placeholder description for OTU_taxonomy dataset.
}
\keyword{datasets}
