% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_seqs}
\alias{test_seqs}
\title{test_seqs}
\format{
An object of class \code{DNAbin} of length 3.
}
\source{
Generated by the Author
}
\usage{
test_seqs
}
\description{
Placeholder description for test_seqs dataset.
}
\keyword{datasets}
