% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training.ASV.table}
\alias{training.ASV.table}
\title{training.ASV.table}
\format{
A data frame with 1909 rows and 4 variables:
\describe{
  \item{Sample_name}{Character vector, last digit includes PCR replicate}
  \item{Locus}{Which locus the sequences are from, numeric}
  \item{Hash}{The sequence found, hashed using sha1 for consistency}
  \item{nReads}{Number of times that hash was found in that sample}
}
}
\source{
Generated by the Author
}
\usage{
training.ASV.table
}
\description{
Placeholder description for training.ASV.table dataset.
}
\keyword{datasets}
